Move from ED25519 to EC, which is support in all (modern) browsers
Changing to EC was painless, however after fixing this new problems appeared.

A major problem here was that the browser implementation of crypto.subtle
enforces that the data input to crypto.subtle.verify() should be an ArrayBuffer,
the node.js implementation is quite happy to eat a string. Addition of
TextEncoders solved this.

Finally, it turns out that since fetch() uses the browser it also uses caching,
so there was a lot of confusion when it didn't load the new publicKey.