Actual support for in-text-signatures
Still at the VERY MUCH prototype stage, currently no considerations of more than
one quote, or of the start/end string appearing more than once on the page.

Additionally, the key is currently expected to be on the server of the page that
is accessed. This is expected to be fixed in next commit, which will hopefully
be a major update of the signatures.